requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2